@@ -1,7 +1,7 @@
 Gem::Specification.new do |s|
   s.platform     = Gem::Platform::RUBY
   s.name         = 'spree_digital'
-  s.version      = '3.3.0'
+  s.version      = '4.2.0.rc1'
   s.summary      = ''
   s.description  = 'Digital download functionality for spree'
   s.authors      = ['funkensturm', 'Michael Bianco']
@@ -11,32 +11,18 @@ Gem::Specification.new do |s|
   s.test_files   = `git ls-files -- {test,spec,features}/*`.split("\n")
   s.require_path = 'lib'
   s.requirements << 'none'
-  s.required_ruby_version = '>= 2.1.0'
+  s.required_ruby_version = '>= 2.5.0'
 
-  spree_version = '>= 3.1.0', '< 4.0'
-  s.add_dependency 'spree_api', spree_version
+  spree_version =  '>= 4.3.0.rc1'
+  s.add_dependency 'spree', spree_version
   s.add_dependency 'spree_backend', spree_version
-  s.add_dependency 'spree_core', spree_version
-  s.add_dependency 'spree_frontend', spree_version
-  s.add_dependency 'spree_extension'
+
+  s.add_runtime_dependency 'spree_extension'
   s.add_dependency 'deface', '~> 1.0'
 
   # test suite
-  s.add_development_dependency 'shoulda-matchers'
-  s.add_development_dependency 'capybara'
-  s.add_development_dependency 'coffee-script'
-  s.add_development_dependency 'factory_bot_rails', '~> 4.7'
-  s.add_development_dependency 'rspec-rails'
-  s.add_development_dependency 'rspec-activemodel-mocks'
-  s.add_development_dependency 'sass-rails'
-  s.add_development_dependency 'sqlite3'
-  s.add_development_dependency 'ffaker'
-  s.add_development_dependency 'database_cleaner'
-  s.add_development_dependency 'simplecov'
-  s.add_development_dependency 'guard-rspec'
   s.add_development_dependency 'rb-fsevent'
-  s.add_development_dependency 'growl'
-  s.add_development_dependency 'appraisal'
-  s.add_development_dependency 'pg', '~> 0.18'
-  s.add_development_dependency 'mysql2', '~> 0.5.1'
+  s.add_development_dependency 'rspec-activemodel-mocks'
+  s.add_development_dependency 'shoulda-matchers'
+  s.add_development_dependency 'spree_dev_tools'
 end
