@@ -3,13 +3,7 @@ module Spree
     belongs_to :variant
     has_many :digital_links, dependent: :destroy
 
-    has_attached_file :attachment, path: ":rails_root/private/digitals/:id/:basename.:extension"
-    do_not_validate_attachment_file_type :attachment
-    validates_attachment_presence :attachment
-
-    if Paperclip::Attachment.default_options[:storage] == :s3
-      attachment_definitions[:attachment][:s3_permissions] = :private
-      attachment_definitions[:attachment][:s3_headers] = { :content_disposition => 'attachment' }
-    end
+    has_one_attached :attachment
+    validates :attachment, attached: true
   end
 end
