@@ -6,10 +6,10 @@ Spree::Core::Engine.routes.draw do
 
     resources :orders do
       member do
-        get :reset_digitals  
+        get :reset_digitals
       end
     end
   end
-  
-  get '/digital/:secret', :to => 'digitals#show', :as => 'digital', :constraints => { :secret => /[a-zA-Z0-9]{30}/ }
+
+  get '/digital/:secret', to: 'digitals#show', as: 'digital', constraints: { secret: /[a-zA-Z0-9]{30}/ }
 end
