@@ -5,14 +5,14 @@ module Spree
   module Calculator::Shipping
     class DigitalDelivery < ShippingCalculator
       preference :amount, :decimal, default: 0
-      preference :currency, :string, default: ->{ Spree::Config[:currency] }
+      preference :currency, :string, default: -> { Spree::Config[:currency] }
 
       def self.description
         Spree.t(:digital_delivery, scope: 'digital')
       end
 
-      def compute_package(package=nil)
-        self.preferred_amount
+      def compute_package(_package = nil)
+        preferred_amount
       end
 
       def available?(package)
