@@ -11,6 +11,7 @@ module Spree
         end
 
         private
+
         def split_by_digital(package)
           digitals = Hash.new { |hash, key| hash[key] = [] }
           package.contents.each do |item|
@@ -21,7 +22,7 @@ module Spree
 
         def hash_to_packages(digitals)
           packages = []
-          digitals.each do |id, contents|
+          digitals.each do |_id, contents|
             packages << build_package(contents)
           end
           packages
