@@ -1,7 +1,7 @@
 require 'spree_api'
 require 'spree_backend'
 require 'spree_core'
-require 'spree_frontend'
 require 'spree_digital/engine'
 require 'sass/rails'
 require 'spree_extension'
+require 'deface'
