@@ -1,23 +1,9 @@
-if ENV["COVERAGE"]
-  # Run Coverage report
-  require 'simplecov'
-  SimpleCov.start do
-    add_group 'Controllers', 'app/controllers'
-    add_group 'Helpers', 'app/helpers'
-    add_group 'Mailers', 'app/mailers'
-    add_group 'Models', 'app/models'
-    add_group 'Views', 'app/views'
-    add_group 'Libraries', 'lib'
-  end
-end
+# Configure Rails Environment
+ENV['RAILS_ENV'] = 'test'
 
-ENV["RAILS_ENV"] = "test"
-require File.expand_path("../dummy/config/environment.rb",  __FILE__)
+require File.expand_path('dummy/config/environment.rb', __dir__)
 
-require 'rspec/rails'
 require 'rspec/active_model/mocks'
-require 'database_cleaner'
-require 'ffaker'
 require 'shoulda-matchers'
 
 Shoulda::Matchers.configure do |config|
@@ -27,43 +13,12 @@ Shoulda::Matchers.configure do |config|
   end
 end
 
-Dir[File.join(File.dirname(__FILE__), "support/**/*.rb")].each {|f| require f }
+require 'spree_dev_tools/rspec/spec_helper'
 
-require 'spree/testing_support/factories'
-require 'spree/testing_support/controller_requests'
-require 'spree/testing_support/url_helpers'
-require 'spree/testing_support/authorization_helpers'
-
-Dir[File.join(File.dirname(__FILE__), "factories/*.rb")].each {|f| require f }
+# Requires supporting ruby files with custom matchers and macros, etc,
+# in spec/support/ and its subdirectories.
+Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].sort.each { |f| require f }
 
 RSpec.configure do |config|
-
-  config.color = true
-  config.disable_monkey_patching!
-  config.raise_errors_for_deprecations!
-  config.infer_spec_type_from_file_location!
-  config.mock_with :rspec
-
-  config.include FactoryBot::Syntax::Methods
-  config.include Spree::TestingSupport::UrlHelpers
-  config.include Spree::TestingSupport::ControllerRequests, :type => :controller
-  config.extend Spree::TestingSupport::AuthorizationHelpers::Controller, :type => :controller
-  config.include SpreeDigital::TestingSupport::Helpers, :type => :controller
-  config.use_transactional_fixtures = false
-
-  config.before(:each) do
-    if RSpec.current_example.metadata[:js]
-      DatabaseCleaner.strategy = :truncation, { except: ['spree_countries', 'spree_zones', 'spree_zone_members', 'spree_states', 'spree_roles'] }
-    else
-      DatabaseCleaner.strategy = :transaction
-    end
-  end
-
-  config.before(:each) do
-    DatabaseCleaner.start
-  end
-
-  config.after(:each) do
-    DatabaseCleaner.clean
-  end
+  config.include SpreeDigital::TestingSupport::Helpers, type: :controller
 end
