@@ -2,7 +2,7 @@ module SpreeDigital
   module TestingSupport
     module Helpers
       def image(filename)
-        File.open(SpreeDigital::Engine.root + "spec/fixtures" + filename)
+        File.open(SpreeDigital::Engine.root + 'spec/fixtures' + filename)
       end
 
       def upload_image(filename)
